% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concaveman-pkg.R
\docType{package}
\name{concaveman-package}
\alias{concaveman-package}
\title{concaveman: A very fast 2D concave hull algorithm.}
\description{
This package is a simple R port (through \href{https://github.com/jeroen/v8}{\code{V8}}) of a \href{https://github.com/mapbox/concaveman}{JavaScript library by Vladimir Agafonkin}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://joelgombin.github.io/concaveman/}
  \item \url{https://github.com/joelgombin/concaveman/}
  \item Report bugs at \url{https://github.com/joelgombin/concaveman/issues}
}

}
\author{
\strong{Maintainer}: Joël Gombin \email{joel.gombin@gmail.com}

Authors:
\itemize{
  \item Ramnath Vaidyanathan
  \item Vladimir Agafonkin
}

Other contributors:
\itemize{
  \item Mapbox [copyright holder]
}

}
\keyword{internal}
