% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concaveman.R
\name{concaveman_rcpp}
\alias{concaveman_rcpp}
\title{concaveman_rcpp}
\usage{
concaveman_rcpp(xy, concavity = 2, length_threshold = 0)
}
\arguments{
\item{xy}{xy}

\item{concavity}{concavity}

\item{length_threshold}{length_threshold}
}
\description{
concaveman_rcpp
}
