% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{points}
\alias{points}
\title{Fixtures data}
\format{an \code{sf} object with a 1000 points. Each of them is part of a group, indicated by variable k (generated by a k-means algorithm).}
\source{
https://github.com/mapbox/concaveman/blob/master/test/fixtures/points-1k.json
}
\usage{
points
}
\description{
This is just a test dataset which comes from the original mapbox library.
}
\keyword{datasets}
